+++ libs/adapter/index.ts
@@ -2,6 +2,7 @@ export * from './event-store-db/event-store-db.event-store';
 export * from './event-store-db/event-store-db.event-subscription';
 
 export * from './in-memory/in-memory.entity-repository';
+export * from './in-memory/in-memory.feature-service';
 
 export * from './mongo-db/mongo-db.entity-repository';
 export * from './mongo-db/mongo-db.helper';