+++ libs/domain/lib/context.ts
@@ -1,0 +1,7 @@
+import { FeatureService } from './feature-service';
+import { LoggerService } from './logger-service';
+
+export interface Context {
+  featureService?: FeatureService
+  loggerService?: LoggerService,
+}