+++ libs/domain/lib/feature-flags.ts
@@ -1,0 +1,4 @@
+import { Feature } from './feature';
+import { Runtime } from './runtime';
+
+export type FeatureFlags = Record<Runtime, Feature>;