+++ libs/domain/lib/feature-service.ts
@@ -1,0 +1,5 @@
+import { Feature } from './feature';
+
+export interface FeatureService {
+  getAll(): Feature;
+}